% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCAData-pkg.R
\docType{internal}
\name{HCAData-pkg}
\alias{HCAData-pkg}
\title{HCAData: Accessing The Datasets Of The Human Cell Atlas in R/Bioconductor}
\description{
\code{HCAData} is an ExperimentHub package which provides access to the single-cell
RNA-seq data from the Human Cell Atlas project (\url{https://www.humancellatlas.org})
}
\references{
If you use the data in this package, please refer to the original sources
(Human Cell Atlas Data Portal, \url{https://preview.data.humancellatlas.org}) as well
(plus the related publications, which will be listed here when they will be out),
which are licensed under a Creative Commons Attribution 4.0 International License.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/federicomarini/HCAData}
}

}
\author{
Federico Marini \email{marinif@uni-mainz.de}
}
